refactor: 통합 구조로 travel-post 디렉토리 재구성
