--- src/features/planned-companion/index.js
@@ -1,13 +1,0 @@
-export * from "./api/plannedApi";
-export * from "./dto/plannedCompanionDto";
-
-// ✅ Components
-export { default as PlannedCompanionCard } from "/src/features/planned-companion/components/PlannedCompanionCard";
-
-// ✅ Modals
-export { default as PlannedCompanionPostModal } from "./modals/PlannedCompanionPostModal";
-export { default as CreatePlannedModal } from "./modals/PlannedCompanionPostModal";
-export { default as UpdatePlannedModal } from "./modals/PlannedCompanionPostModal";
-
-// ✅ Pages
-export { default as PlannedCompanionPage } from "/src/features/planned-companion/pages/PlannedCompanionPage";
