+++ src/features/travel-post/index.js
@@ -1,0 +1,17 @@
+export * from "./api/travelPostApi";
+export * from "./dto/travelPostDto";
+
+// ✅ Components
+export { default as PlannedCompanionCard } from "@/features/travel-post/components/PlannedCompanionCard";
+export { default as LocalCompanionCard } from "@/features/travel-post/components/LocalCompanionCard";
+
+// ✅ Modals
+export { default as PlannedCompanionPostModal } from "@/features/travel-post/modals/PlannedCompanionPostModal";
+export { default as CreatePlannedModal } from "@/features/travel-post/modals/PlannedCompanionPostModal";
+export { default as UpdatePlannedModal } from "@/features/travel-post/modals/PlannedCompanionPostModal";
+export { default as CreateLocalModal } from "@/features/travel-post/modals/LocalCompanionPostModal";
+export { default as UpdateLocalModal } from "@/features/travel-post/modals/LocalCompanionPostModal";
+
+// ✅ Pages
+export { default as PlannedCompanionPage } from "@/features/travel-post/pages/PlannedCompanionPage";
+export { default as LocalCompanionPage } from "@/features/travel-post/pages/LocalCompanionPage";
